server/requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.4
channels==2.4.0
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pyhumps==1.6.1
pylint==2.6.0
pyOpenSSL==19.1.0
pytz==2020.4
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
wrapt==1.12.1
zope.interface==5.2.0
